% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/markdown.R
\name{commonmark}
\alias{commonmark}
\alias{markdown}
\alias{markdown_commonmark}
\alias{markdown_html}
\alias{markdown_man}
\alias{markdown_xml}
\title{Parse and render markdown text}
\usage{
markdown_html(text, sourcepos = FALSE, hardbreaks = FALSE, smart = FALSE,
  normalize = FALSE)

markdown_xml(text, sourcepos = FALSE, hardbreaks = FALSE, smart = FALSE,
  normalize = FALSE)

markdown_man(text, sourcepos = FALSE, hardbreaks = FALSE, smart = FALSE,
  normalize = FALSE)

markdown_commonmark(text, sourcepos = FALSE, hardbreaks = FALSE,
  smart = FALSE, normalize = FALSE, width = 0)
}
\arguments{
\item{text}{Markdown text}

\item{sourcepos}{Include source position attribute}

\item{hardbreaks}{Treat newlines as hard line breaks}

\item{smart}{Use smart punctuation}

\item{normalize}{Consolidate adjacent text nodes}

\item{width}{Specify wrap width (default 0 = nowrap)}
}
\description{
Converts markdown text to several formats using John MacFarlane's \href{https://github.com/jgm/cmark}{cmark}
reference implementation. Output in HTML, groff man, CommonMark, and a custom XML format is supported.
}
\examples{
library(curl)
md <- readLines(curl("https://raw.githubusercontent.com/yihui/knitr/master/NEWS.md"))
html <- markdown_html(md)
xml <- markdown_xml(md)
man <- markdown_man(md)
cm <- markdown_commonmark(md)
}

