% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{extensions}
\alias{extensions}
\alias{list_extensions}
\title{Github CommonMark Extensions}
\usage{
list_extensions()
}
\description{
The Github fork of cmark supports several markdown extensions which provide
features which are not (yet) in the official commonmark spec.
}
\details{
Currently the following extensions are supported:
\itemize{
\item \strong{table} support rendering of tables: \href{https://github.github.com/gfm/#tables-extension-}{gfm-spec section 4.10}
\item \strong{strikethrough} via \verb{~~sometext~~} syntax: \href{https://github.github.com/gfm/#strikethrough-extension-}{gfm-spec section 6.5}
\item \strong{autolink} automatically turn URLs into hyperlinks: \href{https://github.github.com/gfm/#autolinks-extension-}{gfm-spec section 6.9}
\item \strong{tagfilter} blacklist html tags: \code{title} \code{textarea} \code{style} \code{xmp} \code{iframe}
\code{noembed} \code{noframes} \code{script} \code{plaintext}: \href{https://github.github.com/gfm/#disallowed-raw-html-extension-}{gfm-spec section 6.11}
\item \strong{tasklist} turns certain list items into checkboxes: \href{https://github.github.com/gfm/#task-list-items-extension-}{gfm-spec section 5.3}
}

See the full spec for \href{https://github.github.com/gfm/}{GitHub Flavored Markdown}.
}
\examples{
print(list_extensions())
}
