% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/markdown.R
\name{markdown_html}
\alias{markdown_cm}
\alias{markdown_html}
\alias{markdown_man}
\alias{markdown_xml}
\title{Convert markdown to HTML}
\usage{
markdown_html(text, options = 8L)

markdown_xml(text, options = 8L)

markdown_man(text, options = 8L)

markdown_cm(text, options = 8L, width = 100L)
}
\arguments{
\item{text}{markdown text}

\item{options}{integer with options}

\item{width}{width of the commonmark output}
}
\description{
Convert markdown text to HTML using John MacFarlane's \href{https://github.com/jgm/cmark}{cmark}
reference implementation.
}
\examples{
library(curl)
md <- readLines(curl("https://raw.githubusercontent.com/yihui/knitr/master/NEWS.md"))
html <- markdown_html(md)
xml <- markdown_xml(md)
man <- markdown_man(md)
cm <- markdown_cm(md)
}

