% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{extensions}
\alias{extensions}
\alias{list_extensions}
\title{Github CommonMark Extensions}
\usage{
list_extensions()
}
\description{
The Github fork of cmark supports several markdown extensions which provide
features which are not (yet) in the official commonmark spec.
}
\details{
Currently the following extensions are supported:
\itemize{
\item \strong{table} support rendering of tables
\item \strong{strikethough} via \code{~sometext~} syntax
\item \strong{autolink} automatically turn URLs into hyperlinks
\item \strong{tagfilter} blacklist html tags: \code{title} \code{textarea} \code{style} \code{xmp} \code{iframe}
\code{noembed} \code{noframes} \code{script} \code{plaintext}.
}

See also the webpage on \href{https://guides.github.com/features/mastering-markdown/}{GitHub Flavored Markdown}.
}
\examples{
print(list_extensions())
}

